requirements.txt
flask
google-generativeai
python-dotenv
SQLAlchemy
alembic
python-json-logger
eventlet
websocket-client
flask-socketio
python-socketio
flask-admin
flask_sqlalchemy
flask-cors
requests
appdirs
pywin32; sys_platform == 'win32'
Flask-Migrate
neo4j